move S.M.T. string tests, add base class test data
